jvmutil/deps: re-integrate spectral clustering.
this partially reverts b81b910. that commit deleted the spectral
clustering logic, replacing it with logic for computing strongly
connected components. sccs are a good starting place, because they
involve no manual work to untangle deps -- they are "free".

but decomposing a jar into its sccs isn't really desirable -- there
are usually too many. what we really want is a way to decompose a jar
into some number of chunks, no two of which depend on each other.

this commit attempts to do so, by running spectral clustering on the
sccs. the problem is that this can indeed create pairs of chunks that
depend on each other. as a stopgap, the program runs a second round of
scc after the spectral clustering. this at least guarantees that the
final graph is acyclic, but i suspect it's not optimal.