misc_config: maven_deps: bump ktfmt 0.41 -> 0.42.
and reformat the repo with the new version.

the main feature in the new version
(https://github.com/facebook/ktfmt/releases/tag/v0.42) is improved
kdoc formatting. i don't care about that, but unfortunately it removes
some blank lines, which causes the compiled class files to change,
which reruns most tests in the repo. oh well.